fix: split out type exports
