+++ src/lib/configure-security.ts
@@ -1,8 +1,7 @@
+import type { AuthForHAR } from './types';
 import type { OASDocument, SecuritySchemeObject } from 'oas/rmoas.types';
 
 import { isRef } from 'oas/rmoas.types';
-
-export type AuthForHAR = Record<string, string | number | { pass?: string; user?: string }>;
 
 function harValue(type: 'cookies' | 'headers' | 'queryString', value: { name: string; value: string }) {
   if (!value.value) return undefined;
